app/requirements.txt
flask
google-cloud-storage
cryptocode
gunicorn
logging